backend/requirements.txt
streamlit>=1.29.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.1.78
Pillow>=10.0.1
numpy>=1.24.3
pytesseract>=0.3.10
reportlab>=4.0.4
PyMuPDF>=1.23.8
pikepdf>=8.6.0
vosk>=0.3.45
pyttsx3>=2.90
transformers>=4.35.0
sentencepiece>=0.1.99
protobuf>=3.20.3
cryptography>=41.0.7
scikit-image>=0.21.0
scipy>=1.10.1
matplotlib>=3.7.3
easyocr>=1.7.0
pydub>=0.25.1
imageio-ffmpeg>=0.4.9
deep-translator==1.11.4
requests>=2.31.0
edge-tts>=6.1.9
python-dotenv